split monolith into separate files
