+++ src/project_explorer/__main__.py
@@ -1,3 +1,5 @@
-from project_explorer import main
+"""Module main (python -m project_explorer)"""
+
+from project_explorer.viewer import main
 
 main()
